+++ datalogger_jupyterlab/__init__.py
@@ -1,4 +1,5 @@
 """JupyterLab extension configuration."""
 
+
 def _jupyter_labextension_paths():
     return [{"src": "../labextension", "dest": "datalogger-jupyterlab"}]