+++ src/plugins/index.ts
@@ -1,2 +1,3 @@
 export { default as netcdfFileTypePlugin } from "./netcdfFileTypePlugin";
 export { default as dataloggerLoadCodePlugin } from "./dataloggerLoadCodePlugin";
+export { default as pdfPreviewPlugin } from "./pdfPreviewPlugin";