+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as generateLoadCode } from "./generateLoadCode";
 export { default as addToNotebook } from "./addToNotebook";
+export { default as calcPdfPreviewPosition } from "./calcPdfPreviewPosition";